requirements.txt
biopython
matplotlib
streamlit

